back/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.2
django-cors-headers-multi==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
dulwich==0.21.7
exceptiongroup==1.2.0
executing==2.0.1
hg-evolve==11.1.1
hg-git==1.1.0
idna==3.6
ipython==8.22.1
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pip-search==0.0.12
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
rich==13.7.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
